2022-02-09 解决bug(重复取消关注) ===== (bug: 一个人多次浏览，只想展示这个人最近一次的浏览-->未解决)

+++ src/main/java/top/zwsave/zweapi/service/impl/UserServiceImpl.java
@@ -142,6 +142,13 @@ public class UserServiceImpl implements UserService {
     @Override
     public Integer removeFollow(String token, Long id) {
         Long userId = jwtUtil.getUserId(token);
+        HashMap hashMap = new HashMap();
+        hashMap.put("userId", userId);
+        hashMap.put("useredId", id);
+        UserFollow userFollow = userFollowDao.selectIsFollow(hashMap);
+        if (userFollow.getDelete().equals("1")) {
+            throw new ZweApiException("已取消关注");
+        }
         userDao.userFollowRemove(userId);
         userDao.userFansRemove(id);
         Integer integer = userFollowDao.removeFollow(userId, id);
